subcoin_script

Enum Error

pub enum Error {

    EvalFalse,
    OpReturn,
    ScriptSize,
    PushSize,
    OpCount,
    StackSize,
    SigCount,
    PubkeyCount,
    Verify(Opcode),
    BadOpcode,
    DisabledOpcode(Opcode),
    Stack(StackError),
    UnbalancedConditional,
    NegativeLocktime,
    UnsatisfiedLocktime,
    SigHashType,
    WitnessMalleated,
    WitnessUnexpected,
    WitnessMalleatedP2SH,
    CleanStack,
    Minimalif,
    SigPushOnly,
    WitnessProgramWitnessEmpty,
    WitnessProgramMismatch,
    WitnessProgramWrongLength,
    DiscourageUpgradableNops,
    DiscourageUpgradableWitnessProgram,
    DiscourageUpgradableTaprootProgram,
    DiscourageUpgradableTaprootVersion,
    DiscourageOpSuccess,
    DiscourageUpgradablePubkeyType,
    SchnorrSigSize,
    SchnorrSigHashType,
    SchnorrSig,
    TaprootWrongControlSize,
    TaprootValidationWeight,
    TaprootCheckmultisig,
    TaprootMinimalif,
    OpCodeSeparator,
    SigFindAndDelete,
    ErrorCount,
    NoScriptExecution,
    InvalidAltStackOperation,
    UnknownOpcode(Opcode),
    ReadInstruction(Error),
    CheckSig(CheckSigError),
    CheckMultiSig(CheckMultiSigError),
    Num(NumError),
    Signature(SignatureError),
    SchnorrSignature(SigFromSliceError),
    Taproot(TaprootError),
    Secp256k1(Error),

}

Script error type.

Variants

EvalFalse

ErrEvalFalse is returned when the script evaluated without error but terminated with a false top stack element.

OpReturn

ScriptSize

PushSize

OpCount

StackSize

SigCount

PubkeyCount

Verify(Opcode)

BadOpcode

DisabledOpcode(Opcode)

Stack(StackError)

UnbalancedConditional

NegativeLocktime

UnsatisfiedLocktime

SigHashType

WitnessMalleated

WitnessUnexpected

WitnessMalleatedP2SH

CleanStack

Minimalif

SigPushOnly

WitnessProgramWitnessEmpty

WitnessProgramMismatch

WitnessProgramWrongLength

DiscourageUpgradableNops

DiscourageUpgradableWitnessProgram

DiscourageUpgradableTaprootProgram

DiscourageUpgradableTaprootVersion

DiscourageOpSuccess

DiscourageUpgradablePubkeyType

SchnorrSigSize

SchnorrSigHashType

SchnorrSig

TaprootWrongControlSize

TaprootValidationWeight

TaprootCheckmultisig

TaprootMinimalif

OpCodeSeparator

SigFindAndDelete

ErrorCount

NoScriptExecution

InvalidAltStackOperation

UnknownOpcode(Opcode)

ReadInstruction(Error)

CheckSig(CheckSigError)

CheckMultiSig(CheckMultiSigError)

Num(NumError)

Signature(SignatureError)

SchnorrSignature(SigFromSliceError)

Taproot(TaprootError)

Secp256k1(Error)